requirements.txt
kubernetes==3.0.0
google-api-python-client
google-auth-httplib2
